Reduce requests to prevent reaching rate limits
Use the archive Top Songs playlist table to get the list of all
available countries or regions. Then search for Top Songs playlists
based on this list, instead of an exhaustive list of countries
and territories. This should reduce the number of requests, thus
preventing the program from hitting the rate limit.

Silence `conflict_prefer()` to reduce clutter in the console.